#544 refactor code to GraphableEdge.containsPoint()
