+++ kernel/src/main/java/org/cristalise/kernel/lifecycle/instance/Next.java
@@ -22,9 +22,7 @@ package org.cristalise.kernel.lifecycle.instance;
 
 import static org.cristalise.kernel.graph.model.BuiltInEdgeProperties.ALIAS;
 import static org.cristalise.kernel.graph.model.BuiltInEdgeProperties.TYPE;
-
 import org.cristalise.kernel.common.InvalidDataException;
-import org.cristalise.kernel.graph.model.GraphPoint;
 import org.cristalise.kernel.graph.model.GraphableEdge;
 import org.cristalise.kernel.graph.model.GraphableVertex;
 
@@ -67,38 +65,5 @@ public class Next extends GraphableEdge {
         }
 
         throw new InvalidDataException("Terminus Vertex Id:"+getTerminusVertexId()+" was not found in parent:"+getParent().getName());
-    }
-
-    @Override
-    public boolean containsPoint(GraphPoint p) {
-        GraphPoint originPoint = getOriginPoint();
-        GraphPoint terminusPoint = getTerminusPoint();
-        GraphPoint midPoint = new GraphPoint();
-
-        if (("Broken +".equals(getBuiltInProperty(TYPE)))) {
-            midPoint.x = (originPoint.x + terminusPoint.x) / 2;
-            midPoint.y = (originPoint.y + terminusPoint.y) / 2;
-        }
-        else if (("Broken -".equals(getBuiltInProperty(TYPE)))) {
-            boolean arrowOnY = !(originPoint.y - terminusPoint.y < 60 && originPoint.y - terminusPoint.y > -60);
-            midPoint.x = arrowOnY ? terminusPoint.x : (originPoint.x + terminusPoint.x) / 2;
-            midPoint.y = arrowOnY ? (originPoint.y + terminusPoint.y) / 2 : originPoint.y;
-        }
-        else if (("Broken |".equals(getBuiltInProperty(TYPE)))) {
-            boolean arrowOnY = !(originPoint.y - terminusPoint.y < 60 && originPoint.y - terminusPoint.y > -60);
-            midPoint.x = arrowOnY ? originPoint.x : (originPoint.x + terminusPoint.x) / 2;
-            midPoint.y = arrowOnY ? (originPoint.y + terminusPoint.y) / 2 : terminusPoint.y;
-        }
-        else {
-            midPoint.x = originPoint.x + (terminusPoint.x - originPoint.x) / 2;
-            midPoint.y = originPoint.y + (terminusPoint.y - originPoint.y) / 2;
-        }
-
-        int minX = midPoint.x - 10;
-        int minY = midPoint.y - 10;
-        int maxX = midPoint.x + 10;
-        int maxY = midPoint.y + 10;
-
-        return (p.x >= minX) && (p.x <= maxX) && (p.y >= minY) && (p.y <= maxY);
     }
 }
